Reformated the original test with a real one with real mocks
